requirements.txt
# Core dependencies
pydantic==2.5.0
pydantic-settings==2.1.0
aiosmtpd==1.4.4.post2
aiohttp==3.9.1

# Google Cloud
google-cloud-speech==2.21.0

# Celery and Redis
celery==5.3.4
redis==5.0.1

# Email processing
email-validator==2.1.0

# Monitoring and health checks
prometheus-client==0.19.0

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
black==23.11.0
isort==5.12.0
mypy==1.7.1

# HTTP client for health checks
requests==2.31.0